Improved testing for sets
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@6269 e85b7adc-8362-4630-8c63-7469d557c915

+++ test/set/rel-op.cc
@@ -366,7 +366,7 @@ namespace Test { namespace Set {
           {
             IntSetRanges issr(iss);
             unsigned int cardSum = Iter::Ranges::size(issr);
-            if (cardSum != n)
+            if (cardSum != static_cast<unsigned int>(n))
               return false;
             if (withConst) {
               IntSetRanges isr(is);
